add fit module, refactor nmad into spatialstats
