+++ examples/plot_blockwise_coreg.py
@@ -99,5 +99,5 @@ plt.show()
 # %%
 # We can compare the :ref:`spatial_stats_nmad` to validate numerically that there was an improvment:
 
-print(f"Error before: {xdem.spatial_tools.nmad(diff_before):.2f} m")
-print(f"Error after: {xdem.spatial_tools.nmad(diff_after):.2f} m")
+print(f"Error before: {xdem.spatialstats.nmad(diff_before):.2f} m")
+print(f"Error after: {xdem.spatialstats.nmad(diff_after):.2f} m")
