+++ xdem/__init__.py
@@ -1,4 +1,4 @@
-from . import coreg, dem, examples, spatial_tools, spatialstats, volume, filters, terrain
+from . import coreg, dem, examples, spatial_tools, spatialstats, volume, filters, fit, terrain
 from .ddem import dDEM
 from .dem import DEM
 from .demcollection import DEMCollection
